GDBSymbols: Cleanups
