+++ Source/Tools/FEXGDBReader/CMakeLists.txt
@@ -10,3 +10,6 @@ install(TARGETS ${NAME}
 
 target_include_directories(${NAME} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/Source/)
 target_include_directories(${NAME} PRIVATE ${CMAKE_BINARY_DIR}/generated)
+
+# We don't actually link, but this is a nice way to get the include dirs
+target_link_libraries(${NAME} PRIVATE Common)
